salt: Waiting for salt-minion startup state to finish
Salt can not have several states running at the same time. In salt-minion
configuration we have startup state to provision volumes it is done in
`parallel=True` but the state may take time to retrieve pillar
information or stuff like that, so we need to wait before running another
state.
We can't do it directly in the salt minion state as it would always see
itself as a running state.
So we check it in the `metalk8s_saltutil.wait_minions` runner called in
the `deploy_node` orchestrate (where we got issue because trying to
launch state right after restarting the salt-minion)