salt: properly implement `absent` for various objects
For `Role`, `RoleBinding`, `ServiceAccount` and `PodSecurityPolicy` the
states relied upon by the `kubernetes` *renderer* did not actually
exist. This commit implements these properly (also slightly changing how
the renderer handles `absent`, because we want to be able to *remove*
objects without necessarily passing in their whole body/spec), as well
as the 'module' functions required to implement the 'state's.