ci: add command helpers to create build status
Add command helpers to create either SUCCESSFUL
or FAILED status in artifacts, for each step.

These helpers will be used, in a first time, to
be able to retrieve last successful/failed build
for a specific commit through artifacts.

In a second time, we want to improve how we build
in the CI and be able to skip already succeed
steps to avoid potential flakies & improve build
time.

Refs: #1985