ci: add final status to artifact for unit_tests step
Add .final_status file in artifacts containing
the status (SUCCESSFUL or FAILED) of the unit_tests
step

This artifact allows to easily retrieve the status
of a specific step and will, in the near future,
allow to skip the rebuild of an already succeed
step for a specific commit hash.
This will be useful to avoid rebuilding all steps
when encountering a failure due to a flaky
resource or test.

Refs: #1985