scripts/build: Sync all salt modules before updating namespace version
In upgrade and downgrade scripts we annotate the namespace with a
`state.orchestrate_single` so we need to first sync the module and states
of this version to be able to run this.
Since we run the `state.orchestrate_single` in the script directly we
need to sync the states and modules using the version of the script we
can not use the `DEST_VERSION` or the "current node version" we need to
use the script version, so adding a render to add `@@VERSION` in upgrade
and downgrade scripts.