scripts/cli: Refactor logging framework
We extract most of the logging logic into the `metalk8s_cli.log` module.
We also introduce a `verbosity` parameter, along with a `log-level` one,
to allow more fine-grained control of the CLI output.
Commands should now rely on `self.print` to write to `stdout`, and
`self.log` or its variants to log information (which will be written to
both `stderr` and some configured files).