salt: reconfigure salt minion before highstate
We reconfigure & restart the minion before
bringing it to highstate to avoid potential flakies
due to the fact that master is checking for the
highstate job liveness and result every X seconds.
Sometimes it happens during the restart of the
minion, so the master assumes that the job is failed
because minion is unresponsive for few seconds.

Refs: #1952