scripts: use common lib in restore script
Use the common.sh library and remove functions
already moved to this file.
It will avoid doing some changes to each and
every scripts instead of only changing the code
in this library.
It also fixes an issue with repositories configuration
due to the fact we forgot to change this in the restore
script when we changed how we setup package repositories.

Refs: #2033