[WIP] salt/solution: Stop deploying Admin UI cluster-wide
The setup of "environment" Namespaces and all the resources required to
run a Solution instance in it is becoming too complex for it to be
properly managed by the Solution Admin UIs, without mentioning the authz
requirements on users of said Admin UIs to make it work.
Instead, Admin UIs will now be deployed once per environment, making it
easier to keep in sync with the same Operator version.