ui/solutions: Read Solutions info in environments
We add more data about configured Solutions in an environment (through a
'metalk8s-environment' ConfigMap), and the Admin UI for each.
This will be used in the Solutions page to display Environments info.

This changes from the previous approach, as Admin UIs used to be unique
per Solution, not per Solution in each environment.

Refs: #2278