buildchain/mypy: fix typing issue
With this new version of mypy, we have two new errors.

The first one is:

    Argument 1 to "run" has incompatible type "List[object]";
    expected "Union[bytes, str, Sequence[Union[bytes, str, _PathLike[Any]]]]"

This is easily fixed by adding a type annotation on `cmd`.

The second one is:

    Incompatible types in assignment (expression has type "TextIO", variable has type "BinaryIO")
    No overload variant of "write" of "BinaryIO" matches argument type "str"
    Possible overload variants:
        def write(self, s: bytearray) -> int
        def write(self, s: bytes) -> int

This one is due to the reuse of `fp` (a with statement doesn't introduce
a scope, thus the second `fp` is not a new variable, but the old one
reused).
Simply use two different variable to fix this error.

Closes: #2064
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>