ui/ducks/salt: Rework management of Jobs
The initial implementation was focused on Jobs for the deployment of new
Nodes, but this was both limiting (as we cannot re-use it for other
kinds of Salt Jobs) and confusing (actions and sagas in the `nodes` duck
were not actually about Nodes per-se).

We introduce a more generic solution, that decouples the creation,
persistence, and observation of active Jobs from the consumption of the
Jobs' status and result. The `app.salt` duck is re-purposed for this
use-case.

Note that some interesting saga patterns are introduced to cope with the
complexity of this management (buffering of actions to manage a
persisted state in LocalStorage, garbage collection of old persisted
Jobs...).

Also note that, while adapting the NodeDeployment page to this
refactored state (as it is the main consumer of Jobs status and result),
we drastically simplify it: there was no need for an internal state of
the component, as all the required information is available in the Redux
state.