buildchain/targets: add `type: ignore` annotations
PyYAML has type annotations provided by typeshed, but those are
incomplete.

The long term solution would be to improve the type annotations in
typeshed, but for now we will selectively ignore the errors.

Note that we can't simply ignore the whole YAML library (like we do for
`docker` or `doit`) because it *has* type annotations.
Trying to do so results in a mypy warning:

    note: unused 'type: ignore' comment

Refs: #2070
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>