ui/salt/tests: Test the newly introduced sagas
The sagas for managing Salt jobs (in `state.app.salt.jobs`) need good
testing coverage, as they will be the backbone for multiple parts of the
application that rely on those jobs.

Note that a number of constants and functions needed to be exported from
the original duck just for their usage in tests, which is why the
exports are done in this commit.