ui/nodes/test: Test how `fetchNodes` reads Jobs
In order to mark Nodes as "deploying" or not, the `fetchNodes` saga
needs to read the list of Jobs in the state and map each active job to
the corresponding Node (based on node name, compared to Job name, the
latter being expected to be of the form "deploy-node/<node-name>").

Some extra test cases are marked as `todo`, but the nominal path is
covered.