tf: add resources to install the bootstrap node
Add the possiblity to deploy metalk8s on the
bootstrap node depending on an optional variable
`metalk8s_iso`, if defined, it'll try to do it.
Usage:
  terraform apply -auto-approve \
    -var='metalk8s_iso=/path/to/your/iso'

It's also possible to specify the destination of
the ISO `metalk8s_iso_dest` and the mountpoint
`metalk8s_iso_mountpoint` on the remote host.

To only deploy a bootstrap node, it's possible
to do:
  terraform apply -auto-approve \
    -var='metalk8s_iso=/path/to/your/iso' \
    -target=null_resource.metalk8s_run_bootstrap