ci/buildplan: Add build stage status automatically
We introduce a `dsl` module, in which we implement the concept of a
mutating stage decorator. The first application is a decorator that
manages creating and uploading a build status at then end of a stage.

Some core steps were added, namely `ShellCommand` and `Upload`, to
achieve this first application.