ci/buildplan: Add step to retrieve artifacts
Instead of manually running `curl` commands, we create a wrapper that
helps retrieve build artifacts (from the current build only, at the
moment).
This allows to define a "naive" `as_remote` implementation, that simply
downloads the artifacts on the worker and copies them through SCP on the
destination host (most of the time, our remote host won't have access to
the artifacts service).