deploy: Extract Terraform tooling from Eve
We extract our Terraform definition files from the multiple-nodes Eve
worker into a set of re-usable, more generic, modules.
This will help both end-users wanting to try MetalK8s on OpenStack and
developers wanting to setup various test environments.
We plan on using it in our CI in a future enhancement, and keep the
existing Terraform definitions for now.