scripts/iso-manager: override `metalk8s.api_server.kubeconfig` in Pillar
Previous MetalK8s states used during `iso-manager.sh -a` stage expect
the path to the `KubeConfig` to be used by `salt-master` to be available
in the Pillar (`metalk8s.api_server.kubeconfig`). This is removed in a
subsequent commit, replaced by a hard-coded value. As such, we need to
override this value (using the same hard-coded value) in the script,
otherwise importing e.g. a 2.3 ISO on a 2.4.x platform (containing these
changes) will fail.

Keeping this as a separate commit, mainly for documentation purposes.

See: https://github.com/scality/metalk8s/pull/2106