ci: explicitly source `/etc/profile` when running build
For some reason `/etc/profile` isn't exported when `doit` is invoked.
As such, the build is not using the caching HTTP proxy we have in place
in our CI environment, resulting in a lot of unnecessary public traffic
or hitting upstream services.

Fixing by an explicit `source` of the environment file when invoking
`doit.sh`.