tests: get control plane ip from grain
Retrieve control plane IP from grains
metalk8s.control_plane_ip grain instead
of getting the first IP configured on the
controle plane network interface.
With the introduction of the VIP in terraform
the old behavior could lead to retrieving the
VIP address instead of the real IP of the interface.