ci: don't re-run bootstrap in multi-node tests
Re-running bootstrap should not affect anything but the bootstrap node.
Since we run it in single-node CI tests, there's not much use re-running
it in a multi-node environment, especially since we don't actually
*validate* there's no impact on the cluster.

If we were to extend the test-suite to validate this, or when putting
less focus on the single-node environments, we could enable this again.

Note: we now need to ignore 'failure' when no tests are collected,
because there's only a single 'slow' test for now (which we don't want
to execute in the first place). Instead of removing execution of 'slow'
tests altogether in the multi-node scenario (and potentially not adding
the step again when more tests are added), we ignore this specific
failure instead.