salt: Split node deployment procedure
This commit splits node deployment procedure into
a sequence of stages.

Initially, expanding the etcd cluster from 1 to 2 nodes puts
the cluster in a blocked state. This happens when a new member is declared
and not yet started.
If this time to start is too long, liveness probes on the initial etcd Pod will
fail, then it will be restarted and then enter a back-off loop.
Even after the new etcd Pod has been created, the new one won't be able to start,
since the initial is in back-off and will enter a back-off as well.

To prevent the above situation, we pre-pull the etcd image on the target Node,
which ensures the shortest time possible between the new member being registered
and its corresponding Pod being created. Note that this reordering is inspired by
kubespray