contrib/tf: Provision a MetalK8s ISO on Bootstrap
A user can now provide a `metalk8s_iso` Terraform variable to make sure
an ISO is provisioned on the Bootstrap node.
One can use `mode="local"` to `scp` the ISO from one's host, or
`mode="remote"` to let the Bootstrap fetch the ISO from a URL (using
Bastion as proxy if necessary).