build: handle umask for mkdir target
Add the possibility to change the umask when
creating a directory.
This is needed to build the UI as we do it
with a specific user `node`, which needs to
be able to write in the volume which is created
by the local user. So we need to enforce the mode
to 777 on the directory.

Refs: #1595