contrib/tf: Allow specifying a VIP's number
Instead of explicitly picking a VIP's value, we let users set it to a
number defining the position of the VIP within the subnet range
configured.

Note that we bump the required version of Terraform to 0.12.7 and
higher, to be able to use the `regexall` function. In the CI, we switch
to the latest: 0.12.20