contrib/tf: Pick any VIP in subnet
It seems the attribution of port IP addresses is random on some OpenStack
deployments, and forcing a VIP doesn't play well with this behaviour.
To circumvent the issue, we simply let OpenStack choose the IP for our
us. In the future, having full control over which IPs are assigned and
why, to avoid conflicts, would also be interesting.