build: Add custom grafana build image with plugins installed
Updating the prometheus-operator charts to 8.1.2 results in dashboards
with no stats mainly because of missing plugins(grafana-piechart-panel).

In order to have all dashboards with stats, we build our custom grafana
image with the missing plugin installed.

This PR, will be reverted once a new version of the chart is released
since the fix will be included.

Fixes: #2175