build/debian: Fix Saltstack repository URL
We pin an old version of Saltstack to use in our product. This version
is no longer available on the "main" repo domain, but instead under the
"archive" domain. URL thus need to change, though only for Debian (as we
already used the "archive" domain for RedHat).
Note that we also ensure the repo we use is specific to the complete
version we rely on (in this case, 2018.3.4).