kubernetes: Add K8s object creation for persistentvolume and persistentvolumeclaim
Formally, we our python kubernetes module, we are unable to create a pv or pvc.
This is mainly because the models and names for pv and pvc do not exist hence
applying a pvc template will fail.

This is resolved by adding the pv and pvc model to the python k8s list of known
kinds.