build/image: don't touch a file on disk for image
We used to touch a file on disk to keep track of the images we locally
build without saving them on disk (we only keep them in the Docker
registry).
Since 7e47310a8, this is no longer needed, thus we can remove this code
(I've already removed the clean-related code, but forgot this part…).

Refs: #XXXX
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>