build: fix wrong usage of setdefault for task_dep
In those cases, we wanted to add a task_dep to the existing ones (if
any), but the current code was only adding it when there was no existing
task_dep already.

This was noticed when `./doit.sh clean` was sometimes not able to clean
the `_build/root/images` directory and indeed, in same case we had the
following trace:

    […]
    _image_pull:nginx - executing 'Python: bound method ContainerImage.clean of <buildchain.targets.remote_image.RemoteImage object'
    _image_pull:calico-kube-controllers - executing 'Python: bound method ContainerImage.clean of <buildchain.targets.remote_image.RemoteImage object'
    _image_mkdir_root - cannot remove (it is not empty) '/home/eve/workspace/kube_pod-prod-metalk8s-backend-0/build/_build/root/images'
    _build_deb_repositories:salt_ubuntu1804/build_repo - executing 'Python: function DEBRepository.build_repo.<locals>.clean'
    […]
    _package_mkdir_root - removing dir '/home/eve/workspace/kube_pod-prod-metalk8s-backend-0/build/_build/packages'
    _image_build:metalk8s-ui - executing 'Python: bound method ContainerImage.clean of <buildchain.targets.local_image.LocalImage object'
    doc:pdf - executing 'Python: function task_doc.<locals>.clean.<locals>.<lambda>'
    […]

The metalk8s-ui overrides `task_dep`, and thus the deps on `_image_mkdir_root` wasn't set.

The bug only appeared in `_local_image`, but since the pattern was
present in other places, let's fix them all at once before they blow up
one day…

Refs: #XXXX
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>