build/ui: fix _ui_bluid cleanup action
The problem comes from the fact that the files under `UI_BUILD_ROOT` are
created by the user `node` inside the container, and their permissions
is `-rw-r--r--` (only owner can delete them).

Now, this happens to work fine locally because I guess most of people
have UID 1000 for their user (and the user `node` has UID 1000) but in
the CI the user `eve` has UID 1042 and thus it cannot delete those
files…

The cleanest solution is to let the `node` user clean those files, thus
we add a `clean` command to the UI builder entrypoint.

Refs: #2223
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>