docs: add `release` tag
This commit adds a `release` tag to the Sphinx project. Currently, when
set, `todo` elements in the documentation will *not* be rendered. This
is, however, not the default.

A related change is made to pass the `O` environment variable from the
`tox` invocation all the way down to arguments to `sphinx-build`. As
such, one can now run

```
O="-t release" tox -e docs
```

to create 'release' builds of the documentation.