ci: Disable downgrade test from 2.6 to 2.5
Since downgrade of an etcd cluster from 3.4 to 3.3 is impossible (unless
the etcd upgrade was only partial, which is not the case for our test),
we disable this stage from the CI run.

This would require a warning in documentation, with a procedure to
restore an etcd backup for a previous version.