calico: update to Calico 3.12.0
Deployment manifest updated based on upstream from
https://docs.projectcalico.org/v3.12/manifests/calico.yaml.

Note: Calico by default now deploys a CNI configuration which includes
the `bandwidth` plugin. Since this plugin executable is not part of the
`kubernetes-cni` package we deploy, this is not enabled in 'our'
configuration.

Once `kubernetes-cni` would include this binary (or, given where things
are going, `kubelet` includes it), we could enable it.

Another approach would be for our `calico-cni-plugin` to include the
required CNI plugins (as is done by the upstream install mechanism,
copying binaries out of a container), and no longer install
`kubernetes-cni`. However, this would introduce conflicts between
packages installing the same files.