ci: move cypress tests to single-node-rhel step
The cypress tests were done in the downgrade step
but this one has been disabled in 2.6 because
it's not possible to downgrade etcd from this
version.
We move cypress tests to another step to re-enable
them and avoid merging broken UI stuff.
single-node-rhel has been choosen because it's one
of the fastest step, so we're not adding time to
the whole build.

Refs: #2286