build: escape special chars in package URL
"+" character is used as a substitute for
space in URLs.
Normally, the packages remote server should
handle them as plain "+" characters, as there is
no need for spaces in a package name.
A recent change in repo.saltstack.com server
seems to break this behavior, so we can't fetch
following URL anymore:
  http://repo.saltstack.com/apt/ubuntu/18.04/amd64/archive/2018.3.4/pool/main/s/salt/salt-minion_2018.3.4+ds-1_all.deb
To workaround this issue, we escape all special
characters in the path part of the URL, to ensure
they will be well interpreted by the remote
server.

Refs: #2286