salt: Add and deploy an Alertmanager secret
We need to be able to have control over the Alertmanager configurations
which is deployed as a secret using the prometheus-operator charts.

Using the charts, the Alertmanager config is directly base64 encoded
as such jinja substitution during the yaml dump is not possible.

This commit adds a salt sls file that creates the Alertmanager secret