chart: Remove default Alertmanager secret created from charts
Using the helm template engine with Alertmanager secret
(Alertmanager.yaml) does not seem to work as expected because
the templated variables are directly base64 encoded instead
of being replaced during the yaml dump.

We resolve this by disabling the creation of the default secret
for Alertmanager then create a new secret with default values
making it possible for users to edit this configuration setting.

This chart is finally rendered using:

./charts/render.py prometheus-operator --namespace metalk8s-monitoring charts/prometheus-operator.yaml
--service-config grafana metalk8s-grafana-config --service-config prometheus metalk8s-prometheus-config
--service-config alertmanager metalk8s-alertmanager-config charts/prometheus-operator/ >
salt/metalk8s/addons/prometheus-operator/deployed/chart.sls

Closes: #2262