tests: Refactor Cluster and Service Configuration tests to include teardown
Initially, we did not cleanup Cluster and Service Configurations and return them
to their original values. This let to failing tests in the CI since the ConfigMaps
were left with new values and tests ran after upgrade failed.

Now, for each CSC that is modified during the test setup and execution phase
we cleanup and return CSCs to default state

Closes: #2331