build/docker: do not stack `task_error` decorator
We used to stack calls to the `task_error` decorator, one for each error
we wanted to handle.
While this looked nice, it was buggy.

The decorators after the first one were wrapping a decorated function,
not the original one.

Consequently, if we had more than one lever of decorators and the
function raised an error that wasn't handled by the last decorator then
the error was lost:
- the targeted decorator was handling its error as expected and was
  returning a `TaskError`
- but then, the next decorator was seeing a normal return value (a
  TaskError), not an exception and thus `None` was returned => the error
  was lost on the way.

To solve this issue, we've changed the implementation of `task_error`:
we now have a single call that takes a dict mapping an error type to its
handler.

Closes: #2395
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>