storage-operator: handle correctly a deletion request while pending
The way we were handling a deletion while the creation is in progress
was buggy.
We were correctly detecting that we were creating the volume and thus
were postponing the deletion until the creation was done (either with
success or not).
But we were exiting the reconciliation loop too early and never had a
chance to keep continue the volume creation, thus we were forever stuck
in a Pending state (and the deletion was posptone forever as well).

The fix is easy enough though: in such case we don't exit the
reconciliation loop immediatly, to be able to call `deployVolume` and
made some progress until we're stabilized.

Refs: #2409
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>