docs: Read version details from buildchain
We used to hardcode the `version` and `release` variables (and with
improper values, as we want `version` to hold the full version, not just
the {major}.{minor}).

Now, we read this information using the buildchain helpers, and extend
this with a full Git reference and a warning to display on unreleased
builds.

Fixes: #1703