Reinstate "-dev" suffix for 2.4.4
We chose to set "-discontinued" since development/2.4 is not being
actively worked on anymore.
This however has an unusual meaning, which is even false: "the version
2.4.4 will never be released".

Completely removing a version will be achieved by removing the
corresponding development/ branch. For now, since we may need an
upgrade path from 1.x to 2.x, we will keep 2.0 to 2.4 open.