ci: does not rely on worker_uuid anymore
In the CI context we were using worker_uuid property
as a unique identifier prefix to all terraform resources
we created, to avoid collision between these resources.
It happens that this UUID is not really unique and sometimes
we have uncleaned resources on the MetalK8s scality.cloud CI
tenant, so we end up having collision between old resources
and new ones, this cause builds to fail.

We now use the build number + the stage name as a prefix.
The build number should be unique accross any build and the
stage name will be used to easily retrieve the resources of
a specific stage of a build.