ui/oidc: Request 'groups' scope with id_token
In the scopes we request from Dex when retrieving and id_token to then
authenticate to K8s API, we didn't include the 'groups' optional scope.
This meant that apiserver couldn't retrieve the 'groups' claim from the
generated token, hence the authenticated user wouldn't be able to assume
roles based on Group-based (Cluster)RoleBindings.

We add this scope to the `redux-oidc` manager configuration, and this
resolves the issue.

Fixes: #2526