salt/drain: Handle retryable errors when evicting
The Eviction API sends '429 Too Many Requests' when a requested eviction
can't be applied due to some disruption budget. This way, the client can
wait and retry later.
We didn't handle this error in our implementation, hence rolling
upgrades were failing as soon as we hit this situation.
We add support for this, and also move the "timeout" scope to the whole
eviction process (since we now may be stuck at the eviction creation).

Fixes: #2530