tooling: add tooling to extract predefined Prometheus alerting and recording rules
For documentation purposes, we need to find a way to generate, save and document
the existing Prometheus alerting and recording rules.

This commit adds a python script that queries the Prometheus API and returns Prometheus
rules.

The alerting & recording rules are generated using:

```
$ ./tools/rule_extractor/rule_extractor.py -i 172.21.254.3 -p 8443 -t rules

Where -i is the Control-plane ingress Ip
and -p is the Control-plane ingress Port
	or

$ ./tools/rule_extractor/rule_extractor.py -t rules

```