test: validate that the deployed Prometheus alerting rules match the default alerting rules
We need to be sure that the deployed Prometheus alerting rules match the
default alerting rules at all times.

This test can typically fail if there is a recent Prometheus-operator chart
upgrade that brings along  new Prometheus alerting rules.

We need to be notified when alerting rules differ(when this test fails) so that Developers
can follow the new procedure to update the default rules and include the required
documentation for the newly added alerting rule