storage-operator/salt: use a handle instead of an ID for jobs
Until now, we only had a single Salt async job per step (one to prepare,
one to finalize), thus we knew which job was running by knowing the step
we were in.

This is gonna change, as we want to call `disk.dump` in an async fashion
and thus we will have two async jobs for the prepare step.
In order to know which is which, we now carry the job name alongside its
ID and bundle this up into a JobHandle.

This commit only introduces the JobHandle and uses it instead of the
JobID.
Making `disk.dump` async will be done in another commit.

Refs: #2493
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>