tests: add pytest-bdd scenario for prometheus customization
This test is to ensure that Prometheus CSC is working
fine and we can customize Prometheus alert rules.

This scenario changes a threshold in the Prometheus
CSC ConfigMap, then re-deploys Prometheus customized
rules through Salt and checks that the value
is correctly applied on the PrometheusRule object in
Kubernetes.
Finally, the CSC is restored and the Salt formula re-run
to go back to the original state of the PrometheusRule
object.

Refs: #2448