salt: Add 'metalk8s:debug' pillar
In order to provide a simple toggle to users at the BootstrapConfig
level, we setup this `debug` option, which will be used across all
formulas to setup debug logging/sockets where applicable.
This is only a very basic approach for now (either everything is in
debug mode, or not), which should be improved on later.
However, it should provide some form of debugging help which we don't
have at the moment.