salt: Use debug pillar in containerd states
Note that we extend the systemd service drop-in to modify the
`ExecStart` directive, which allows us to add some command-line
arguments (if they were not exposed through environment vars).
We use this new flexibility to add a `--log-level`, and set it to
`debug` if `metalk8s:debug` pillar is enabled.