salt: update Solutions module to handle manifest
We know use a `manifest.yaml` instead of the
`product.txt` + the `config.yaml` to retrieve a
Solution information such as the name, the version, ...
Also rename the kind from `SolutionConfig` to `Solution`
to avoid confusion with `SolutionsConfiguration` and
because it's not a configuration file.
Note that the `manifest.yaml` file is now mandatory
unlike the old `config.yaml`.

Refs: #2422