example: generate manifest when building Solution
We now generate the new `manifest.yaml` instead of
the `product.txt` + `config.yaml` files with all the
content of both files in a new format as follows:

This change is done to simplify Solution archives,
with only a single file to describe the whole
Solution.

Also, the `config.yaml` was poorly named as it is not
a configuration file, but more a manifest.
`SolutionConfig` has been renamed to `Solution` to avoid
ambiguity with `SolutionsConfiguration` kind.

Refs: #2422