storage-operator/salt: interpret retcode depending on Function
The retcode interpretation is different between Salt states and Salt
modules.
For a Salt states we have case where the job failed because it couldn't
run at all (retcode 1), which is different from "the job ran and failed"
(retcode 2).
But that's only for states, for other kind of jobs we have a retcode 1
for "the job ran and failed".

Closes: #2598
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>