eve: split post-merge in two stages
We were building previous version and
example-solution in a first step, then
triggering tests on solutions and upgrade
/downgrade of the platform as a second step.
If, for example, the building of example solution
failed, we weren't doing the upgrade/downgrade
tests, while they're not related.

We now split the post-merge in two distinct
parts, one for lifecycle (build previous version,
upgrade, downgrade, restore) and one for solution
(build solution and deploy it). So if one of the
two builds fails, the tests for the other are
still run.

This has been done to avoid losing all the
benefits of the post-merge when a single component
is failing.

It also reduces time needed for the post-merge
tests as the Example Solution build can take ages.