workaround kubernetes yum repository bug
See https://github.com/kubernetes/kubernetes/issues/92242
for details.

The right thing to do would be to backport changes (related
to how we fetch packages) done in development/2.4, but
it seems to be some work to do as there is plenty of stuff
to backport in buildchain. For now we need a quick
workaround.