salt-api: Rework Basic auth handling
Instead of relying on the strange assumption that the /login 'password'
argument is actually the full 'user:password' base64-encoded token, we
adjust the storage-operator and metalk8s-ui to pass in this password in
cleartext.

This commit also updates the `_auth_basic` handler to rely on the same
SelfSubjectAccessReview API as the  `_groups_basic` handler did.