salt: add pillar refresh to volumes.prepared
We now call `metalk8s_volumes.device_info` instead of `disk.dump`.
Unlike the later, this new function relies on the pillar content and we
have a problem here because calling a Salt module doesn't refresh the
pillar.

That being said, the state we call before (volumes.prepared) does
automatially refresh the pillar because it works fine and it does look
up the volume from the pillar.
But it seems this pillar isn't kept around, since calling the module
right after fails.
If we explicitely refresh the pillar during our state, then it's kept
around and our module can reuse the fetched pillar: that's why we add a
refresh_pillar call to our state.

Now, in the unlikely case that we have a crash/restart after the state
execution BUT before executing the module, we won't have access to a
cached pillar: in such case we have a last-chance try/except that
refresh the pillar from inside the module.

I'm not super happy with the current approach…, but I don't see a better
one:
- using a state is not possible (because I want to return data)
- calling `saltutil.refresh_pillar` from the operator will complicate
  the code a bit and won't even solve the "I have a crash/reboot/restart
  just before calling my module and thus no cached pillar!" situation.

Refs: #2421
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>