storage-operator/controller: use a volatile cache for device info
Instead of persisting device information into the status (which is
hackish), we now store them in a hashmap.
It will be good enough for most of the cases, and in the worst case (we
crash after getting the device info but before creating the PV), we can
simply reschedule the Salt call to fetch the device information.

Refs: #2421
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>