salt: Ability to provide several CIDRs for control and workload networks
In metalk8s we may have node not in the same L2 networks so that some
routing was needed to communicate between node and it's perfectly find
just we need to allow multiple CIDRs for workload plane and control
plane networks.
Now in bootstrap config you can provide a list of CIDRs for this 2
networks.
Bump apiVersion of bootstrap config as the layout of the file changed