salt/volume: fix `metalk8s_volumes.device_name`
Since Salt error handling is perfectible™, we have to roll-up our own
solution.
In case of error, the old code was producing something like that:

    {
      "return": [ {
        "bootstrap": "The minion function caused an exception: <insert traceback here>"
      }]
    }

Which was wrongly interpreted as a device name…

Raising `CommandExecutionError` instead of `Exception` is a bit better:

    {
      "return": [ {
        "bootstrap": "ERROR: device `the-bad-path` not found"
      }]
    }

but still useless because you can't tell the difference between a
success and a failure without inspecting/parsing the returned string…

One last trick is to set `__context__['retcode']` to a nonzero value
before raising, and then you get:
    {
      "return": [{
        "bootstrap": {
           "success": false,
           "error": "device `the-bad-path` not found"
        }
      }]
    }

which is better BUT you still get a string (instead of a dict) in the
success case (so the client code need to handle the deserialization of
both a dict and a string).

At that point, it just became easier and simpler to change the return
value of our module's function to return a dict(success, result) (which
is homogeneous in both passing and failing case) and update the parsing
code in the operator accordingly.

Closes: #2728
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>