build,script,salt: Pin salt version in bootstrap and restore
In bootstrap and restore we install salt-minion on the node so we need
to pin the salt version we want to install and not just install the one
yum will find depending on various things (like available versions,
priority on repos, ...).
Also set the priority to 1 on all our repos so that our repositories has
priority over others if the `yum-plugin-priorities` package is
installed.
NOTE: If `yum-plugin-priorities` is installed and priority is set on one
repos then other repos with the same package will just be ignored
(e.g.: epel installed with priority=98 we cannot install `salt-3000.3`
as salt-3000.3 is excluded because a repo with some priority has the
package in a different version)

Fixes: #2704