salt/tests: Define custom Kubernetes API mock
This mock allows to perform simple CRUD operations on an in-memory
database following the same principles as K8s API, and is built to be
used through `metalk8s_kubernetes` mocked methods (when testing other
execution modules relying on them).
In addition, this mock can be configured to evolve through "time", using
a list of "events", which are particularly useful to emulate the
reconciliation loops happening in a K8s cluster, while running our tests
much faster than real-time.

See: #2266