salt/tests: Add unit tests for `metalk8s_drain`
This module is slightly different from others, as most logic lives
within the `Drain` class, which is responsible for performing
side-effects. Refactoring this to provide better isolation would be too
long, so we test it as a black box and simply mock K8s API (and the
time).

In the YAML test cases, lists of Pods, ReplicaSets, etc. are stored
under the `datasets` key, which is then read by the mocks when setting
up a test case.

See: #2266