tests: Factorize volume status checking
We extract common logic used to wait for a Volume to reach a given state
and add it to the VolumeClient utility.
We also extract the "test_volume" Given fixture, so that it uses this
new helper, and is now accessible to other test modules. We do not use
it in the Volume tests however, as these rely on finer control (in
particular, they store the resulting object in a `context` fixture).