salt/volume: fix calls to `metalk8s_volumes.device_name`
Since 72324c2, the `device_name` function now return a dict with a
success status instead of raising an error.
The problem is that several callsites where still expecting a string as
return value (and an exception in case of error).

Instead of patching each callsite to translate the error code into
exception, this commit introduce an helper that restore the old
behavior.

Refs: #2784
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>