salt/dex: Handle conversion of ServiceConfig
When applying metalk8s.addons.dex.deployed.service-configuration, if the
existing ConfigMap is in the old format, we will automatically convert
it to the new /v1alpha2 format.
When downgrading, we will automatically convert to the old format.

Note that when performing this conversion, we ensure the previous
configuration is backed up in the same ConfigMap, under the
'previous-config.yaml' key, just in case.

See: #2781