salt/dex: Store managed config in defaults
We previously saved most of the configuration data in the user-editable
ConfigMap. However, some of this configuration needs to be managed by us
(namely, anything related to the control-plane Ingress controller), and
letting this in the ConfigMap prevents any change on our side (which is
likely to happen on upgrades or Bootstrap restore).

The defaults, previously defined in a static YAML file, are now
requiring Jinja rendering, so we slightly adjust the loading behaviour
(using `salt.slsutil.renderer` and relying on its parsing of the
"shebang" in the source file).

See: #2781