salt, charts: Use slsutil.renderer everywhere
Instead of the previous import_yaml Jinja directive, we now rely on the
slsutil.renderer execution method, which is more generic and will let us
change defaults for CSC without changing the rendered charts or any
formula consuming it.
Note that we update the render script to use this behaviour, and
re-render all affected charts accordingly.