doc: Add section to disable the static user store
This user store is not perfect, and doesn't benefit from the same
features as a full-fledged identity management system would. As such,
it may make sense for users of MetalK8s to disable it, and only use Dex
as an intermediary between their identity provider(s) and MetalK8s
services.

See: #2781