go/operator-lib/config: Rework error handling
Following advice from https://blog.golang.org/go1.13-errors, we make
sure that errors raised from this package are easy to handle:
- they either wrap an error the caller can manage (e.g.
  `os.ErrNotExist`) or a publicly exposed `ErrValidation`
- the validation error messages are consistent and always include the
  key(s) which caused the validation error

Note that we don't want to explicitly wrap the internal errors (like
`errMissingField`) because they are not public, hence we don't expect
this API to be stable and don't want to test against them.
On the other hand, the user-facing messages are part of the public API,
which is why we still test them.

See: #2837