salt: no longer user pillar to get available solutions
During Salt master installation, we need to retrieve
the list of Solutions ISO to mount into the Salt
master container, to do that we we're relying on the
`metalk8s_solutions` external pillar to retrieve the
list of Solutions.
The issue is that, we need to upgrade the master
using the `--local` option of Salt to avoid having
it unresponsive during the execution of the state
(because the Salt master container can be replaced
if its configuration change).
The wrong side of using this option is that we can't
access external pillar computed by the Salt master.
We were ending up with a configuration change during
the upgrade orchestration, even we if the Salt master
was upgraded prior to the upgrade, which broke the
whole upgrade because the master became unresponsive.
So, we now rely directly on the metalk8s_solutions
module, which anyway makes more sense since these
solutions ISO are local to the bootstrap host.

Refs: #2832