ui/cypress: Emulate an already logged in user
Instead of relying on an actual OIDC provider to manage the
authentication dance, we inject a JWT token in the browser's
LocalStorage to make the `oidc-client` believe we have an already logged
in user.
This logic is captured in a new command called `login`, which will make
it easy for various spec files to use it.

See: #2878