ui/cypress: Add integration tests for navbar
We add a complete suite of tests to cover the various functionalities of
the navbar (various links, language selection, expected title, user
management) in isolation from other parts of the application.

Since this provides extended coverage of this component, we remove
redundant E2E tests used to validate language selection and link to
documentation.

See: #2878