ci/ui-unit-tests: Pre-install npm dependencies
We were running `npm ci` in a step, whereas this makes sense to be
cached in the worker image as much as possible (especially since we give
the full `ui/` folder as a build context).
We rely on `npm` cache to store as much as possible, and use it during
the CI run to make it faster (`--prefer-offline`).

See: #2878